fix(fte): keep half a post instead of truncating it away
Approving a 2.5 FTE staffing request opened a position for 2: headcount was
an integer column and the approval did int(needed_fte). The net-open figure
rounded the same way. Across sixteen hotels that adds up to real headcount
nobody budgeted away.

headcount is a float now, the approval carries the FTE across unchanged, and
net_open keeps two decimals instead of rounding to whole posts. PostgreSQL
needs the column widened, so the migration list gained an ALTER; SQLite
already stores it by type affinity.

On screen a raw 2.5 would read wrong in Turkish, so a small fte() helper
prints whole numbers plain and fractions with a comma — "FTE Testi (x2,5)",
total "51,5". The position form now steps by 0.5 and is labelled FTE rather
than "kişi", since half a person is not a thing.

+++ backend/models.py
@@ -92,7 +92,7 @@ class Position(Base):
     # v4 new fields
     is_active = Column(Boolean, default=True)
     location = Column(String, nullable=True)
-    headcount = Column(Integer, default=1)
+    headcount = Column(Float, default=1)   # FTE - yarım kadro (0,5) mümkün
     priority = Column(String, default="Orta")
     hiring_manager = Column(String, nullable=True)
     target_date = Column(String, nullable=True)
